Combined program and Heat map program
The generatetrackandbtf.m program combines all the files that were
written previously except for the visualization program. generateheatmap
uses a 2-D histogram to create a heatmap so as to visualise how
frequently a monkey visits a particular location in the cage. The plan
is to generate these heatmaps and matrix of interaction for every day
seperately