Program to write gml files
writegmlfile.m generates the gml files that can be run on gephi to get
graphs. Also made some changes to the visulaiztion and heatmap programs.
The interaction is calculated based on 3d distance between monkeys,
interaction orbs have a constant radius, the visualization has square
axes. the heat map can be generated between a specified range of
coordinates (in this case the cage is between 0 and 3m in either
direction so generate the heat map between those coordinates)